+++ api/src/main/java/de/maxhenkel/voicechat/api/opus/OpusEncoderMode.java
@@ -1,0 +1,8 @@
+package de.maxhenkel.voicechat.api.opus;
+
+/**
+ * The different Opus encoder modes
+ */
+public enum OpusEncoderMode {
+    VOIP, AUDIO, RESTRICTED_LOWDELAY;
+}