+++ common/src/main/java/de/maxhenkel/voicechat/plugins/impl/opus/OpusEncoderImpl.java
@@ -1,6 +1,7 @@
 package de.maxhenkel.voicechat.plugins.impl.opus;
 
 import de.maxhenkel.voicechat.api.opus.OpusEncoder;
+import de.maxhenkel.voicechat.api.opus.OpusEncoderMode;
 import de.maxhenkel.voicechat.config.ServerConfig;
 import de.maxhenkel.voicechat.voice.client.SoundManager;
 
@@ -15,8 +16,18 @@ public class OpusEncoderImpl implements OpusEncoder {
     }
 
     @Nullable
-    public static OpusEncoderImpl create() {
-        de.maxhenkel.voicechat.voice.client.OpusEncoder encoder = de.maxhenkel.voicechat.voice.client.OpusEncoder.createEncoder(SoundManager.SAMPLE_RATE, SoundManager.FRAME_SIZE, 1024, ServerConfig.Codec.VOIP.getOpusValue());
+    public static OpusEncoderImpl create(@Nullable OpusEncoderMode mode) {
+        int application = ServerConfig.Codec.VOIP.getOpusValue();
+
+        if (mode != null) {
+            application = switch (mode) {
+                case VOIP -> ServerConfig.Codec.VOIP.getOpusValue();
+                case AUDIO -> ServerConfig.Codec.AUDIO.getOpusValue();
+                case RESTRICTED_LOWDELAY -> ServerConfig.Codec.RESTRICTED_LOWDELAY.getOpusValue();
+            };
+        }
+
+        de.maxhenkel.voicechat.voice.client.OpusEncoder encoder = de.maxhenkel.voicechat.voice.client.OpusEncoder.createEncoder(SoundManager.SAMPLE_RATE, SoundManager.FRAME_SIZE, 1024, application);
         if (encoder == null) {
             return null;
         }